Update patch for HEAD bb3bd6f & use suckless standard naming convention (#3)
I tried to reduce the general footprint of the patch by getting rid of
the helper functions in drw.c. That also had the positive side effect of
adding utf8 support to the status bar as we're now using drw_text to
draw the text.
To replace the behavior of x_set_color i create an additional scheme
which is used to store the foreground color of whatever is being drawn.